requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.5.4
fastapi==0.110.0
uvicorn==0.29.0
